Revert "sched/walt: Improve the scheduler"
This reverts commit 44a9fe8ad93e05e8be5bfb2fc839249c6f993592.

+++ include/linux/sched.h
@@ -847,7 +847,7 @@ struct task_struct {
 	struct list_head grp_list;
 	u64 cpu_cycles;
 	bool misfit;
-	u32 unfilter;
+	u8 unfilter;
 #endif
 
 #ifdef CONFIG_CGROUP_SCHED
